♻️refactor: UI 관리 로직 contextAPI -> zustand 로 변경
