+++ src/components/uiContext/UiContext.tsx
@@ -1,103 +1,31 @@
 "use client";
 
-import React, { useCallback, useMemo } from "react";
+import React, { useMemo } from "react";
 import { ThemeProvider } from "styled-components";
 
 import { useLocalStorage } from "@lib/hooks/useLocalStorage";
+import useModalStore from "@/stores/ui/useModalStore";
+import { MODAL_VIEWS } from "@/stores/ui/types/modalType";
+import { darkTheme, lightTheme } from "@/styles/theme";
 import { Modal } from "@components/Modal";
 import { TestModal } from "@components/modalViews";
-import { MODAL_ACTION, MODAL_VIEWS } from "./types/modalType";
-import { darkTheme, lightTheme } from "@/styles/theme";
+
 import LogInitConfirmModal from "../modalViews/LogInitConfirmModal/LogInitConfirmModal";
 
-export interface State {
-  displayModal: boolean;
-  modalView: MODAL_VIEWS;
-  modalProps: any | null;
-}
-
-const initialState: State = {
-  displayModal: false,
-  modalView: "INIT_VIEW",
-  modalProps: null,
-};
-
-type Action = MODAL_ACTION & {};
-
-export const UIContext = React.createContext<State | any>(initialState);
-
-UIContext.displayName = "UIContext";
-
-function uiReducer(state: State, action: Action) {
-  switch (action.type) {
-    case "OPEN_MODAL": {
-      return {
-        ...state,
-        modalProps: action.props,
-        displayModal: true,
-      };
-    }
-    case "CLOSE_MODAL": {
-      return {
-        ...state,
-        modalProps: null,
-        displayModal: false,
-      };
-    }
-    case "SET_MODAL_VIEW": {
-      return {
-        ...state,
-        modalProps: null,
-        modalView: action.view,
-      };
-    }
-  }
-}
-
-export const UIProvider: React.FC = (props) => {
-  const [state, dispatch] = React.useReducer(uiReducer, initialState);
-
-  const openModal = useCallback(
-    (props?: any) => dispatch({ type: "OPEN_MODAL", props }),
-    [dispatch],
-  );
-
-  const closeModal = useCallback(() => dispatch({ type: "CLOSE_MODAL" }), [dispatch]);
-
-  const setModalView = useCallback(
-    (view: MODAL_VIEWS) => dispatch({ type: "SET_MODAL_VIEW", view }),
-    [dispatch],
-  );
-
-  const value = useMemo(
-    () => ({
-      ...state,
-      openModal,
-      closeModal,
-      setModalView,
-    }),
-    [state],
-  );
-
-  return (
-    <UIContext.Provider
-      value={value}
-      {...props}
-    />
-  );
-};
-
-interface IUIContext extends State {
-  openModal: any;
-  closeModal: any;
-  setModalView: (view: MODAL_VIEWS) => void;
-}
-
 export const useUI = () => {
-  const context: IUIContext = React.useContext(UIContext);
-  if (context === undefined) {
-    throw new Error(`useUI must be used within a UIProvider`);
-  }
+  const { displayModal, modalView, modalProps, setModalView, openModal, closeModal } =
+    useModalStore();
+
+  const context = {
+    // modal
+    displayModal,
+    modalView,
+    modalProps,
+    openModal: (props?: any) => openModal(props),
+    closeModal: () => closeModal(),
+    setModalView: (view: MODAL_VIEWS) => setModalView(view),
+  };
+
   return context;
 };
 
@@ -141,8 +69,6 @@ export const ManagedUIContext = ({ children }: { children: any }) => {
   }, [localTheme]);
 
   return (
-    <ThemeProvider theme={themeMode === "light" ? lightTheme : darkTheme}>
-      <UIProvider>{children}</UIProvider>
-    </ThemeProvider>
+    <ThemeProvider theme={themeMode === "light" ? lightTheme : darkTheme}>{children}</ThemeProvider>
   );
 };
