--- src/components/uiContext/types/modalType.ts
@@ -1,14 +1,0 @@
-export type MODAL_VIEWS = "INIT_VIEW" | "LOG_INIT_CONFIRM";
-
-export type MODAL_ACTION =
-  | {
-      type: "OPEN_MODAL";
-      props?: any;
-    }
-  | {
-      type: "CLOSE_MODAL";
-    }
-  | {
-      type: "SET_MODAL_VIEW";
-      view: MODAL_VIEWS;
-    };
